Make mod compatible with Android 13
For some reason lsposed is not able to find the "show" and "hide" methods of "KeyguardDisplayManager" in Android 13.

Probably this issue is the same reported at https://github.com/LSPosed/LSPosed/issues/2088

However, after some tests I see the method "updateDisplays" is still hookable, so I changed the code to rely on it, then it's working again.

I think this is a fragile solution so it's likely some future changes in Android break it again.